ISSUE #1916 - Remove old viewers
