+++ frontend/modules/tree/tree.sagas.ts
@@ -19,7 +19,7 @@ import { all, call, put, select, take, takeLatest } from 'redux-saga/effects';
 
 import { VIEWER_EVENTS } from '../../constants/viewer';
 import * as API from '../../services/api';
-import { Viewer } from '../../services/viewer/viewerService/viewerController';
+import { Viewer } from '../../services/viewer/viewer';
 import { DialogActions } from '../dialog';
 import { GroupsActions } from '../groups';
 import { dispatch } from '../store';
