+++ frontend/routes/viewerGui/components/groups/groups.container.ts
@@ -36,7 +36,7 @@ import {
 } from '../../../../modules/groups';
 import { selectSettings } from '../../../../modules/model';
 import { selectIsModelLoaded } from '../../../../modules/viewerGui';
-import { withViewer } from '../../../../services/viewer/viewerService/viewerController';
+import { withViewer } from '../../../../services/viewer/viewer';
 import { Groups } from './groups.component';
 
 const mapStateToProps = createStructuredSelector({
