+++ frontend/routes/viewerGui/components/pinButton/pinButton.container.ts
@@ -19,7 +19,7 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { createStructuredSelector } from 'reselect';
 import { MeasureActions } from '../../../../modules/measure';
-import { withViewer } from '../../../../services/viewer/viewerService/viewerController';
+import { withViewer } from '../../../../services/viewer/viewer';
 import { PinButton } from './pinButton.component';
 
 const mapStateToProps = createStructuredSelector({});
