+++ frontend/routes/viewerGui/components/views/views.component.tsx
@@ -26,7 +26,7 @@ import { VIEWER_EVENTS } from '../../../../constants/viewer';
 import { VIEWER_PANELS } from '../../../../constants/viewerGui';
 import { renderWhenTrue } from '../../../../helpers/rendering';
 import { IViewpointsComponentState } from '../../../../modules/viewpoints/viewpoints.redux';
-import { Viewer } from '../../../../services/viewer/viewerService/viewerController';
+import { Viewer } from '../../../../services/viewer/viewer';
 import { PanelBarActions } from '../panelBarActions';
 import { ViewerPanelButton, ViewerPanelFooter } from '../viewerPanel/viewerPanel.styles';
 import { PresetViews } from './components/presetViews/presetViews.component';
