+++ frontend/services/viewer/viewer.tsx
@@ -1,69 +1,124 @@
+/**
+ **  Copyright (C) 2020 3D Repo Ltd
+ **
+ **  This program is free software= you can redistribute it and/or modify
+ **  it under the terms of the GNU Affero General Public License as
+ **  published by the Free Software Foundation, either version 3 of the
+ **  License, or (at your option) any later version.
+ **
+ **  This program is distributed in the hope that it will be useful,
+ **  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ **  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ **  GNU Affero General Public License for more details.
+ **
+ **  You should have received a copy of the GNU Affero General Public License
+ **  along with this program.  If not, see <http=//www.gnu.org/licenses/>.
+ **/
+
+import EventEmitter from 'eventemitter3';
 import React from 'react';
 
 import { IS_DEVELOPMENT } from '../../constants/environment';
-import { INITIAL_HELICOPTER_SPEED, VIEWER_PIN_MODE } from '../../constants/viewer';
-import { Viewer as ViewerInstance } from '../../globals/viewer';
+import {
+	VIEWER_EVENTS,
+	VIEWER_NAV_MODES,
+} from '../../constants/viewer';
+import { UnityUtil } from '../../globals/unity-util';
 import { DialogActions } from '../../modules/dialog';
 import { dispatch, getState } from '../../modules/store';
 import { selectMemory } from '../../modules/viewer';
+import { PIN_COLORS } from '../../styles';
 import { clientConfigService } from '../clientConfig';
 import { MultiSelect } from './multiSelect';
 
+const UNITY_LOADER_PATH = 'unity/Build/UnityLoader.js';
+
+declare const Module;
+
+interface IViewerConstructor {
+	name?: string;
+}
+
 export class ViewerService {
-	public viewer = null;
-	public pin: any;
-	public newPinId: string;
+	public element: HTMLElement;
+	public name: string;
+	public viewer: HTMLElement;
+	public currentNavMode = null;
+	public units = 'm';
+	public convertToM = 1.0;
+	public initialized = false;
+	public measureMode = false;
+	public modelString = null;
+	public divId = 'unityViewer';
+	private numClips = 0;
+	private stats: boolean = false;
+	private emitter = new EventEmitter();
 	public initialised: {
 		promise: Promise<any>;
 		resolve: () => void;
 		reject: () => void;
 	};
 
-	private mode = VIEWER_PIN_MODE.NORMAL;
-	private pinData: any;
+	public fullscreen: boolean;
+	public pinDropMode: boolean;
+	public unityLoaderReady: boolean;
+	public unityLoaderScript: HTMLScriptElement;
+	public settings: any;
+	public options: any;
+	public plugins: any;
 
-	private stats: boolean = false;
-	public helicopterSpeed = INITIAL_HELICOPTER_SPEED;
-	private container;
+	public constructor({ name = 'viewer', ...config}: IViewerConstructor) {
+		this.name = name;
 
-	constructor() {
-		this.newPinId = 'newPinId';
-		this.pinData = null;
+		this.unityLoaderReady = false;
+		this.pinDropMode = false;
 
-		this.pin = {
-			pinDropMode: false
-		};
+		this.viewer = document.createElement('div');
+		this.viewer.className = 'viewer';
 	}
 
-	get isPinMode() {
-		return this.mode === VIEWER_PIN_MODE.PIN;
-	}
-
-	public async isViewerReady() {
-		await this.initialised.promise;
-	}
-
-	public async isModelReady() {
-		await this.viewer.isModelLoaded();
-	}
+	/**
+	 * Setup Instance
+	 */
 
 	public get hasInstance() {
-		return !!this.viewer;
+		return !!this.element;
 	}
+
+	public setupInstance = (container) => {
+		this.element = container;
+		UnityUtil.init(this.handleUnityError, this.onUnityProgress, this.onModelProgress);
+		UnityUtil.hideProgressBar();
+
+		const unityHolder = document.createElement('div');
+		unityHolder.className = 'emscripten';
+		unityHolder.setAttribute('id', this.divId);
+		unityHolder.removeAttribute('style');
+		unityHolder.setAttribute('width', '100%');
+		unityHolder.setAttribute('height', '100%');
+		unityHolder.setAttribute('tabindex', '1'); // You need this for unityHolder to register keyboard events
+		unityHolder.setAttribute('oncontextmenu', 'event.preventDefault()');
+
+		unityHolder.onmousedown = () => {
+			return false;
+		};
+
+		unityHolder.style['pointer-events'] = 'all';
+
+		this.element.appendChild(this.viewer);
+		this.viewer.appendChild(unityHolder);
+
+		this.unityLoaderScript = document.createElement('script');
+	}
+
+	/**
+	 * Initialization
+	 */
 
 	public get memory() {
 		const MAX_MEMORY = 2130706432; // The maximum memory Unity can allocate
 		const assignedMemory = selectMemory(getState()) * 1024 * 1024; // Memory is in Mb.
 		return Math.min(assignedMemory, MAX_MEMORY);
-	}
-
-	public setupInstance = (container, name = 'viewer') => {
-		this.container = container;
-		this.viewer = new ViewerInstance({
-			name,
-			container: this.container,
-			onError: this.handleUnityError
-		});
 	}
 
 	public init = async () => {
@@ -72,397 +127,176 @@ export class ViewerService {
 		}
 
 		this.setInitialisePromise();
+		UnityUtil.viewer = this;
 
-		this.viewer.setUnity();
 		try {
-			await this.viewer.insertUnityLoader(this.memory);
+			await this.insertUnityLoader(this.memory);
 
-			setTimeout(() => {
-				this.viewer.init({
+			(async () => {
+				await this.initUnity({
 					getAPI: {
 						hostNames: clientConfigService.apiUrls.all
 					},
 					showAll: true
 				});
-			}, 1000);
+
+				this.initialised.resolve();
+			})();
 		} catch (error) {
 			console.error('Error while initialising Unity script: ', error);
 		}
 	}
 
-	public setInitialisePromise() {
-		const initialised = {} as any;
-		initialised.promise = new Promise((resolve, reject) => {
-			initialised.resolve = resolve;
-			initialised.reject = reject;
+	public initUnity(options) {
+		return new Promise((resolve, reject) => {
+			if (this.initialized) {
+				resolve();
+			}
+
+			UnityUtil.setAPIHost(options.getAPI);
+
+			// Set option param from viewerDirective
+			this.options = options;
+			this.emit(VIEWER_EVENTS.VIEWER_INIT);
+			document.body.style.cursor = 'wait';
+
+			// Shouldn't need this, but for something it is not being recognised from unitySettings!
+			Module.errorhandler = UnityUtil.onUnityError;
+
+			if (this.options && this.options.plugins) {
+				this.plugins = this.options.plugins;
+				Object.keys(this.plugins).forEach((key) => {
+					if (this.plugins[key].initCallback) {
+						this.plugins[key].initCallback(this);
+					}
+				});
+			}
+
+			this.setNavMode(VIEWER_NAV_MODES.TURNTABLE, false);
+
+			UnityUtil.onReady().then(() => {
+				this.initialized = true;
+				this.emit(VIEWER_EVENTS.UNITY_READY, {
+					model: this.modelString,
+					name: this.name
+				});
+				resolve();
+			}).catch((error) => {
+				this.emit(VIEWER_EVENTS.VIEWER_INIT, error);
+				console.error('UnityUtil.onReady failed: ', error);
+				reject(error);
+			});
 		});
-		this.initialised = initialised;
 	}
 
-	public destroy = () => {
-		this.viewer.destroy();
-		this.container = null;
-	}
-
-	public async updateViewerSettings(settings) {
-		await this.isViewerReady();
-		return this.viewer.updateSettings(settings);
-	}
-
-	public async updateClippingPlanes(params) {
-		await this.isViewerReady();
-		return this.viewer.updateClippingPlanes(
-			params.clippingPlanes,
-			params.account,
-			params.model
-		);
-	}
-
-	public getNumPlanes() {
-		if (this.viewer) {
-			return this.viewer.getNumPlanes();
+	public insertUnityLoader(memory) {
+		if (document.querySelector('.unity-loader')) {
+			return Promise.resolve();
 		}
-	}
 
-	public centreToPoint(params: any) {
-		if (this.viewer) {
-			this.viewer.centreToPoint(params);
-		}
-	}
+		return new Promise((resolve, reject) => {
+			this.unityLoaderScript.addEventListener ('load', () => {
+				(async () => {
+					await UnityUtil.loadUnity(this.divId, undefined, memory);
+					console.debug('Loaded UnityLoader.js succesfully');
+					resolve();
+				})();
+			}, false);
+			this.unityLoaderScript.addEventListener ('error', (error) => {
+				console.error('Error loading UnityLoader.js', error);
+				reject('Error loading UnityLoader.js');
+			}, false);
 
-	public on = (...args) => {
+			// Event handlers MUST come first before setting src
+			this.unityLoaderScript.src = UNITY_LOADER_PATH;
 
-		this.viewer.on(...args);
-	}
-
-	public once = (...args) => {
-
-		this.viewer.once(...args);
-	}
-
-	public off = (...args) => {
-
-		this.viewer.off(...args);
-	}
-
-	public async topView() {
-		await this.isModelReady();
-		return this.viewer.topView();
-	}
-
-	public async bottomView() {
-		await this.isModelReady();
-		return this.viewer.bottomView();
-	}
-
-	public async frontView() {
-		await this.isModelReady();
-		return this.viewer.frontView();
-	}
-
-	public async backView() {
-		await this.isModelReady();
-		return this.viewer.backView();
-	}
-
-	public async leftView() {
-		await this.isModelReady();
-		return this.viewer.leftView();
-	}
-
-	public async rightView() {
-		await this.isModelReady();
-		return this.viewer.rightView();
-	}
-
-	public async setCamera(params) {
-		await this.isModelReady();
-
-		return this.viewer.setCamera(
-			params.position,
-			params.view_dir,
-			params.up,
-			params.look_at,
-			params.animate !== undefined ? params.animate : true,
-			params.rollerCoasterMode,
-			params.account,
-			params.model
-		);
-	}
-
-	public async goToDefaultViewpoint() {
-		await this.isViewerReady();
-		return this.viewer.showAll();
-	}
-
-	public removeUnsavedPin() {
-		this.removePin({ id: this.newPinId });
-		this.setPin({ data: null });
-	}
-
-	public clearHighlights = async () => {
-		await this.isViewerReady();
-		this.viewer.clearHighlights();
-	}
-
-	public async getCurrentViewpoint({ teamspace, model }) {
-		if (this.viewer) {
-			return await this.viewer.getCurrentViewpointInfo(teamspace, model);
-
-		}
-		return {};
-	}
-
-	public getScreenshot = async () => {
-		await this.isViewerReady();
-		return await this.viewer.getScreenshot();
+			// This kicks off the actual loading of Unity
+			this.unityLoaderScript.setAttribute('class', 'unity-loader');
+			document.body.appendChild(this.unityLoaderScript);
+		});
 	}
 
 	/**
-	 * Pins
+	 * Emitter
 	 */
-	public setPinDropMode(on: boolean) {
-		this.viewer.setPinDropMode(on);
 
-		if (on) {
-			MultiSelect.toggleAreaSelect(false);
-		}
+	public on = (event, fn, ...args) => {
+		this.emitter.on(event, fn, ...args);
 	}
 
-	public setPin(data) {
-		this.pinData = data;
+	public once = (event, fn, ...args) => {
+		this.emitter.once(event, fn, ...args);
 	}
 
-	public getPinData(): any {
-		return this.pinData;
+	public off = (event, ...args) => {
+		this.emitter.off(event, ...args);
 	}
 
-	public async addPin(params) {
-		await this.isViewerReady();
-		this.viewer.addPin(
-			params.account,
-			params.model,
-			params.id,
-			params.type,
-			params.pickedPos,
-			params.pickedNorm,
-			params.colours,
-			params.viewpoint
-		);
-		if (params.isSelected) {
-			this.viewer.selectPin(params.id);
-		}
+	public emit = (event, ...args) => {
+		this.emitter.emit(event, ...args);
 	}
 
-	public async removePin(params) {
-		await this.isViewerReady();
-		this.viewer.removePin(params.id);
+	public removeAllListeners() {
+		this.emitter.removeAllListeners();
 	}
 
-	public async changePinColor({ id, colours }) {
-		await this.isViewerReady();
-		this.viewer.changePinColours(id, colours);
+	public pickPointEvent(pointInfo) {
+		// User clicked a mesh
+		this.emit(VIEWER_EVENTS.PICK_POINT, {
+			id : pointInfo.id,
+			normal : pointInfo.normal,
+			position: pointInfo.position,
+			screenPos : pointInfo.mousePos,
+			selectColour : PIN_COLORS.YELLOW,
+		});
 	}
 
 	/**
-	 * Measure
+	 * Helpers
 	 */
 
-	public async activateMeasure() {
-		await this.isViewerReady();
-		this.viewer.setMeasureMode(true);
+	public async isModelLoaded() {
+		await UnityUtil.onReady();
+		return UnityUtil.onLoaded();
 	}
 
-	public async disableMeasure() {
-		await this.isViewerReady();
-		this.viewer.setMeasureMode(false);
+	public getDefaultHighlightColor() {
+		return UnityUtil.defaultHighlightColor;
+	}
+
+	public getScreenshot() {
+		return UnityUtil.requestScreenShot();
+	}
+
+	public getCurrentViewpointInfo(account, model) {
+		return UnityUtil.requestViewpoint(account, model);
+	}
+
+	public hideHiddenByDefaultObjects() {
+		UnityUtil.hideHiddenByDefaultObjects();
+	}
+
+	public showHiddenByDefaultObjects() {
+		UnityUtil.showHiddenByDefaultObjects();
+	}
+
+	public showCoordView() {
+		UnityUtil.showCoordView();
+	}
+
+	public hideCoordView() {
+		UnityUtil.hideCoordView();
+	}
+
+	public getUnityObjectsStatus(account, model) {
+		return UnityUtil.getObjectsStatus(account, model);
 	}
 
 	/**
-	 * Compare
+	 * Handlers
 	 */
 
-	public diffToolSetAsComparator(teamspace: string, model: string) {
-		this.viewer.diffToolSetAsComparator(teamspace, model);
-	}
-
-	public diffToolLoadComparator(teamspace: string, model: string, revision: string) {
-		return this.viewer.diffToolLoadComparator(teamspace, model, revision);
-	}
-
-	public async diffToolEnableWithClashMode() {
-		await this.isViewerReady();
-		return this.viewer.diffToolEnableWithClashMode();
-	}
-
-	public async diffToolEnableWithDiffMode() {
-		await this.isViewerReady();
-		return this.viewer.diffToolEnableWithDiffMode();
-	}
-
-	public diffToolDisableAndClear = async () => {
-		await this.isViewerReady();
-		return this.viewer.diffToolDisableAndClear();
-	}
-
-	public diffToolShowBaseModel = async () => {
-		await this.isViewerReady();
-		return this.viewer.diffToolShowBaseModel();
-	}
-
-	public diffToolShowComparatorModel = async () => {
-		await this.isViewerReady();
-		return this.viewer.diffToolShowComparatorModel();
-	}
-
-	public diffToolDiffView = async () => {
-		await this.isViewerReady();
-		return this.viewer.diffToolDiffView();
-	}
-
-	public async getObjectsStatus({ teamspace, model } = { teamspace: '', model: '' }) {
-		await this.isViewerReady();
-		return this.viewer.getObjectsStatus(teamspace, model);
-	}
-
-	public async highlightObjects(params) {
-		await this.isViewerReady();
-		this.viewer.highlightObjects(
-			params.account,
-			params.model,
-			params.id ? [params.id] : params.ids,
-			params.zoom,
-			params.colour,
-			params.multi,
-			params.forceReHighlight
-		);
-	}
-
-	public async unhighlightObjects(params) {
-		await this.isViewerReady();
-		this.viewer.unhighlightObjects(
-			params.account,
-			params.model,
-			params.id ? [params.id] : params.ids
-		);
-	}
-
-	public getNavMode() {
-		if (this.viewer) {
-			return this.viewer.currentNavMode;
-		}
-	}
-
-	public getMultiSelectMode() {
-		if (this.viewer) {
-			return this.viewer.multiSelectMode;
-		}
-		return false;
-	}
-
-	public setMultiSelectMode(value) {
-		if (this.viewer) {
-			this.viewer.setMultiSelectMode(value);
-		}
-	}
-
-	public async startAreaSelect() {
-		await this.isViewerReady();
-		this.viewer.startAreaSelect();
-	}
-
-	public async stopAreaSelect() {
-		await this.isViewerReady();
-		this.viewer.stopAreaSelect();
-	}
-
-	public async getDefaultHighlightColor() {
-		await this.isViewerReady();
-		return this.viewer.getDefaultHighlightColor();
-	}
-
-	public async overrideMeshColor(account, model, meshIDs, color) {
-		await this.isViewerReady();
-		this.viewer.overrideMeshColor(account, model, meshIDs, color);
-	}
-
-	public async resetMeshColor(account, model, meshIDs) {
-		await this.isViewerReady();
-		this.viewer.resetMeshColor(account, model, meshIDs);
-	}
-
-	public async goToExtent() {
-		await this.isViewerReady();
-		this.viewer.showAll();
-	}
-
-	public async setNavigationMode(mode) {
-		await this.isViewerReady();
-		this.viewer.setNavMode(mode);
-	}
-
-	public async helicopterSpeedDown() {
-		await this.isViewerReady();
-		this.viewer.helicopterSpeedDown();
-	}
-
-	public async helicopterSpeedUp() {
-		await this.isViewerReady();
-		this.viewer.helicopterSpeedUp();
-	}
-
-	public async helicopterSpeedReset() {
-		await this.isViewerReady();
-		this.viewer.helicopterSpeedReset();
-	}
-
-	public async startClip(isSingle) {
-		await this.isViewerReady();
-
-		if (isSingle) {
-			this.viewer.startSingleClip();
-		} else {
-			this.viewer.startBoxClip();
-		}
-	}
-
-	public async startClipEdit() {
-		await this.isViewerReady();
-		this.viewer.startClipEdit();
-	}
-
-	public async stopClipEdit() {
-		await this.isViewerReady();
-		this.viewer.stopClipEdit();
-	}
-
-	public async showCoordView() {
-		await this.isViewerReady();
-		this.viewer.showCoordView();
-	}
-
-	public async hideCoordView() {
-		await this.isViewerReady();
-		this.viewer.hideCoordView();
-	}
-
-	public async switchObjectVisibility(teamspace, modelId, objectIds, visibility) {
-		await this.isViewerReady();
-		this.viewer.switchObjectVisibility(teamspace, modelId, objectIds, visibility);
-	}
-
-	public async hideHiddenByDefaultObjects() {
-		await this.isViewerReady();
-		this.viewer.hideHiddenByDefaultObjects();
-	}
-
-	public async showHiddenByDefaultObjects() {
-		await this.isViewerReady();
-		this.viewer.showHiddenByDefaultObjects();
-	}
-
-	public handleUnityError = (message: string, reload: boolean, isUnity: boolean) => {
-
+	private handleUnityError = (message: string, reload: boolean, isUnity: boolean) => {
 		let errorType = '3D Repo Error';
 
 		if (isUnity) {
@@ -482,11 +316,322 @@ export class ViewerService {
 		console.error('Unity errored and user canceled reload', message);
 	}
 
-	public reset() {
-		if (this.viewer) {
-			this.disableMeasure();
-			this.viewer.reset();
+	private onUnityProgress = (progress) => {
+		if (progress === 1) {
+			this.emit(VIEWER_EVENTS.VIEWER_INIT_SUCCESS, progress);
+		} else {
+			this.emit(VIEWER_EVENTS.VIEWER_INIT_PROGRESS, progress);
 		}
+	}
+
+	public onModelProgress = (progress) => {
+		this.emit(VIEWER_EVENTS.MODEL_LOADING_PROGRESS, progress);
+	}
+
+	/**
+	 * Destroying
+	 */
+
+	public async destroy() {
+		UnityUtil.reset();
+		this.initialized = false;
+		this.removeAllListeners();
+		this.setPinDropMode(false);
+		await this.disableMeasure();
+	}
+
+	/**
+	 * Selecting
+	 */
+
+	public get canSelect() {
+		return !this.pinDropMode && !this.measureMode;
+	}
+
+	public objectSelected(pointInfo) {
+		if (this.canSelect) {
+			if (pointInfo.id) {
+				if (pointInfo.pin) {
+					// User clicked a pin
+					this.emit(VIEWER_EVENTS.CLICK_PIN, {
+						id: pointInfo.id
+					});
+				} else {
+					this.emit(VIEWER_EVENTS.OBJECT_SELECTED, {
+						account: pointInfo.database,
+						id: pointInfo.id,
+						model: pointInfo.model,
+						source: 'viewer'
+					});
+				}
+			} else {
+				this.emit(VIEWER_EVENTS.BACKGROUND_SELECTED);
+			}
+		} else {
+			if (!pointInfo.id) {
+				this.emit(VIEWER_EVENTS.BACKGROUND_SELECTED_PIN_MODE);
+			}
+		}
+	}
+
+	public objectsSelected(nodes) {
+		if (this.canSelect) {
+			if (nodes) {
+				this.emit(VIEWER_EVENTS.MULTI_OBJECTS_SELECTED, { selectedNodes: nodes });
+			} else {
+				this.emit(VIEWER_EVENTS.BACKGROUND_SELECTED);
+			}
+		}
+	}
+
+	/**
+	 * Measure
+	 */
+
+	public async activateMeasure() {
+		await this.isViewerReady();
+		this.setMeasureMode(true);
+	}
+
+	public async disableMeasure() {
+		await this.isViewerReady();
+		this.setMeasureMode(false);
+	}
+
+	public setMeasureMode(on: boolean) {
+		this.measureMode = on;
+		if (on === true) {
+			UnityUtil.enableMeasuringTool();
+		} else {
+			UnityUtil.disableMeasuringTool();
+		}
+	}
+
+	/**
+	 * Highlight
+	 */
+
+	public get canHighlight() {
+		return this.initialized && !this.pinDropMode && !this.measureMode;
+	}
+
+	public async highlightObjects(
+		{ account, model, colour, multi, forceReHighlight, ...params }:
+		{ account: string, model: string, colour, multi: boolean, forceReHighlight: boolean, id?: string, ids?: string[]}) {
+		if (this.canHighlight) {
+			const ids = params.id ? [params.id] : params.ids;
+			if (ids) {
+				const uniqueIds = Array.from(new Set(ids));
+				if (uniqueIds.length) {
+					// @ts-ignore
+					await UnityUtil.highlightObjects(account, model, uniqueIds, colour, multi, forceReHighlight);
+					this.emit(VIEWER_EVENTS.HIGHLIGHT_OBJECTS, {account, model, uniqueIds });
+					return;
+				}
+			}
+
+			this.clearHighlights();
+		}
+	}
+
+	public clearHighlights() {
+		UnityUtil.clearHighlights();
+		this.emit(VIEWER_EVENTS.CLEAR_HIGHLIGHT_OBJECTS, {});
+	}
+
+	public unhighlightObjects({ account, model, ...params }) {
+		const ids = params.id ? [params.id] : params.ids;
+		if (ids) {
+			const uniqueIds = Array.from(new Set(ids));
+			if (uniqueIds.length) {
+				// @ts-ignore
+				UnityUtil.unhighlightObjects(account, model, uniqueIds);
+				this.emit(VIEWER_EVENTS.UNHIGHLIGHT_OBJECTS, {account, model, uniqueIds });
+				return;
+			}
+		}
+	}
+
+	/**
+	 * Settings
+	 */
+
+	public updateViewerSettings(settings) {
+		if (settings) {
+			this.settings = settings;
+			if (this.settings.properties && this.settings.properties.unit) {
+				this.setUnits(this.settings.properties.unit);
+			}
+		}
+	}
+
+	public setUnits(units) {
+		this.units = units;
+
+		if (units === 'mm') {
+			this.convertToM = 0.001;
+		} else if (units === 'ft') {
+			this.convertToM = 0.0032;
+		}
+
+		// Set the units in unity for the measure tool
+		if (this.units) {
+			UnityUtil.setUnits(this.units);
+		}
+	}
+
+	/**
+	 * Navigation
+	 */
+
+	public helicopterSpeedDown() {
+		UnityUtil.helicopterSpeedDown();
+	}
+
+	public helicopterSpeedUp() {
+		UnityUtil.helicopterSpeedUp();
+	}
+
+	public helicopterSpeedReset() {
+		UnityUtil.helicopterSpeedReset();
+	}
+
+	public setNavigation(mode) {
+		UnityUtil.setNavigation(mode);
+	}
+
+	public setNavigationMode(mode) {
+		this.setNavMode(mode, false);
+	}
+
+	public setNavMode(mode, force) {
+		if (this.currentNavMode !== mode || force) {
+			this.currentNavMode = mode;
+			this.setNavigation(mode);
+		}
+	}
+
+	public navMethodChanged(newNavMode) {
+		this.currentNavMode = newNavMode;
+	}
+
+	public get getNavMode() {
+		return this.currentNavMode;
+	}
+
+	/**
+	 * Fullscreen
+	 */
+
+	public switchFullScreen() {
+		if (!this.fullscreen) {
+			if (this.viewer.hasOwnProperty('mozRequestFullScreen')) {
+			} else if (this.viewer.webkitRequestFullscreen) {
+				this.viewer.webkitRequestFullscreen();
+			}
+
+			this.fullscreen = true;
+		} else {
+			if (document.webkitCancelFullScreen) {
+				document.webkitCancelFullScreen();
+			}
+
+			this.fullscreen = false;
+		}
+	}
+
+	/**
+	 * Pins
+	 */
+
+	public addPin = async ({ id, type, pickedPos: position, norm, colours, isSelected }) => {
+		await this.isViewerReady();
+		await this.isModelLoaded();
+		if (type === 'risk') {
+			UnityUtil.dropRiskPin(id, position, norm, colours);
+		} else {
+			UnityUtil.dropIssuePin(id, position, norm, colours);
+		}
+		if (isSelected) {
+			UnityUtil.selectPin(id);
+		}
+	}
+
+	public removePin({id}) {
+		UnityUtil.removePin(id);
+	}
+
+	public changePinColor({id, colours}) {
+		UnityUtil.changePinColour(id, colours);
+	}
+
+	public setPinDropMode(on: boolean) {
+		this.pinDropMode = on;
+
+		if (on) {
+			MultiSelect.toggleAreaSelect(false);
+		}
+	}
+
+	/**
+	 * Diffs
+	 */
+
+	public diffToolSetAsComparator(account: string, model: string) {
+		UnityUtil.diffToolSetAsComparator(account, model);
+	}
+
+	public diffToolLoadComparator(account: string, model: string, revision: string) {
+		return UnityUtil.diffToolLoadComparator(account, model, revision);
+	}
+
+	public diffToolEnableWithDiffMode() {
+		UnityUtil.diffToolEnableWithDiffMode();
+	}
+
+	public diffToolEnableWithClashMode() {
+		UnityUtil.diffToolEnableWithClashMode();
+	}
+
+	public diffToolDisableAndClear() {
+		UnityUtil.diffToolDisableAndClear();
+	}
+
+	public diffToolShowBaseModel() {
+		UnityUtil.diffToolShowBaseModel();
+	}
+
+	public diffToolShowComparatorModel() {
+		UnityUtil.diffToolShowComparatorModel();
+	}
+
+	public diffToolDiffView() {
+		UnityUtil.diffToolDiffView();
+	}
+
+	/**
+	 * Load
+	 */
+
+	public setInitialisePromise() {
+		const initialised = {} as any;
+		initialised.promise = new Promise((resolve, reject) => {
+			initialised.resolve = resolve;
+			initialised.reject = reject;
+		});
+		this.initialised = initialised;
+	}
+
+	public async isViewerReady() {
+		return await this.initialised.promise;
+	}
+
+	/**
+	 * Load Model
+	 */
+
+	public async isModelReady() {
+		return await this.isModelLoaded();
 	}
 
 	public async loadViewerModel(teamspace, model, branch, revision) {
@@ -494,59 +639,73 @@ export class ViewerService {
 			console.error('Teamspace, model, branch or revision was not defined!', teamspace, model, branch, revision);
 			await Promise.reject('Teamspace, model, branch or revision was not defined!');
 		} else {
-			await this.viewer.loadModel(teamspace, model, branch, revision);
+			await this.loadNewModel(teamspace, model, branch, revision);
 			this.initialised.resolve();
 		}
 	}
 
-	public async zoomToHighlightedMeshes() {
-		await this.isModelReady();
-		this.viewer.zoomToHighlightedMeshes();
+	public async loadNewModel(account, model, branch, revision) {
+		await UnityUtil.onReady();
+		this.emit(VIEWER_EVENTS.MODEL_LOADING_START);
+		document.body.style.cursor = 'wait';
+
+		await UnityUtil.loadModel(account, model, branch, revision);
+
+		await UnityUtil.onLoaded().then((bbox) => {
+			document.body.style.cursor = 'initial';
+
+			this.emit(VIEWER_EVENTS.MODEL_LOADED, 1);
+			this.emit(VIEWER_EVENTS.BBOX_READY, bbox);
+		}).catch((error) => {
+			document.body.style.cursor = 'initial';
+			if (error !== 'cancel') {
+				console.error('Unity error loading model= ', error);
+			}
+		});
+
+		return UnityUtil.onLoading();
 	}
 
-	public async zoomToObjects(meshes) {
-		await this.isModelReady();
-		this.viewer.zoomToObjects(meshes);
+	/**
+	 * Zooms
+	 */
+
+	public zoomToHighlightedMeshes() {
+		UnityUtil.zoomToHighlightedMeshes();
 	}
 
-	public setShadows = (type: string) => {
-		switch (type) {
-			case 'soft':
-				this.viewer.enableSoftShadows();
-				break;
-			case 'hard':
-				this.viewer.enableHardShadows();
-				break;
-			case 'none':
-				this.viewer.disableShadows();
-				break;
+	public zoomToObjects(meshes) {
+		UnityUtil.zoomToObjects(meshes);
+	}
+
+	/**
+	 * Mesh Color
+	 */
+
+	public overrideMeshColor(account, model, meshIDs, color) {
+		UnityUtil.overrideMeshColor(account, model, meshIDs, color);
+
+		if (color.length > 3) {
+			UnityUtil.overrideMeshOpacity(account, model, meshIDs, color[3]);
 		}
 	}
 
-	public setStats = (val: boolean = false) => {
-		if (val !== this.stats) {
-			this.viewer.toggleStats();
-			this.stats = val;
-		}
+	public resetMeshColor(account, model, meshIDs) {
+		UnityUtil.resetMeshOpacity(account, model, meshIDs);
+		UnityUtil.resetMeshColor(account, model, meshIDs);
 	}
 
-	public setNearPlane = (nearplane: number) => {
-		if (nearplane === undefined) { return; }
-		this.viewer.setDefaultNearPlane(nearplane);
-	}
-
-	public setFarPlaneSamplingPoints = (farplaneSample: number) => {
-		if (farplaneSample === undefined) { return; }
-		this.viewer.setFarPlaneSampleSize(farplaneSample);
-	}
+	/**
+	 * Setters
+	 */
 
 	public setFarPlaneAlgorithm = (algorithm: string) => {
 		switch (algorithm) {
 			case 'box':
-				this.viewer.useBoundingBoxFarPlaneAlgorithm();
+				UnityUtil.useBoundingBoxFarPlaneAlgorithm();
 				break;
 			case 'sphere':
-				this.viewer.useBoundingSphereFarPlaneAlgorithm();
+				UnityUtil.useBoundingSphereFarPlaneAlgorithm();
 				break;
 		}
 	}
@@ -554,81 +713,257 @@ export class ViewerService {
 	public setShading = (shading: string) => {
 		switch (shading) {
 			case 'standard':
-				this.viewer.setRenderingQualityDefault();
+				UnityUtil.setRenderingQualityDefault();
 				break;
 			case 'architectural':
-				this.viewer.setRenderingQualityHigh();
+				UnityUtil.setRenderingQualityHigh();
 				break;
 		}
 	}
 
 	public setXray = (xray: boolean) => {
 		if (xray) {
-			this.viewer.setXRayHighlightOn();
+			UnityUtil.setXRayHighlightOn();
 		} else {
-			this.viewer.setXRayHighlightOff();
+			UnityUtil.setXRayHighlightOff();
 		}
-	}
-
-	public async resetMapSources() {
-		await this.isViewerReady();
-		this.viewer.resetMapSources();
-	}
-
-	public async addMapSource(source) {
-		await this.isViewerReady();
-		this.viewer.addMapSource(source);
-	}
-
-	public async removeMapSource(source) {
-		await this.isViewerReady();
-		this.viewer.removeMapSource(source);
-	}
-
-	public async mapInitialise(surveyPoints) {
-		await this.isViewerReady();
-		this.viewer.mapInitialise(surveyPoints);
-	}
-
-	public async mapStart() {
-		await this.isViewerReady();
-		this.viewer.mapStart();
-	}
-
-	public async mapStop() {
-		await this.isViewerReady();
-		this.viewer.mapStop();
-	}
-
-	public setNumCacheThreads(value: number) {
-		if (value === undefined) { return; }
-		this.viewer.setNumCacheThreads(value);
 	}
 
 	public setMaxShadowDistance(value: number) {
 		if (value === undefined) { return; }
-		this.viewer.setMaxShadowDistance(value);
+		UnityUtil.setMaxShadowDistance(value);
+	}
+
+	public setNumCacheThreads(value: number) {
+		if (value === undefined) { return; }
+		UnityUtil.setNumCacheThreads(value);
+	}
+
+	public setNearPlane = (nearplane: number) => {
+		if (nearplane === undefined) { return; }
+		UnityUtil.setDefaultNearPlane(nearplane);
+	}
+
+	public setFarPlaneSamplingPoints = (farplaneSample: number) => {
+		if (farplaneSample === undefined) { return; }
+		UnityUtil.setFarPlaneSampleSize(farplaneSample);
+	}
+
+	public setStats = (val: boolean = false) => {
+		if (val !== this.stats) {
+			UnityUtil.toggleStats();
+			this.stats = val;
+		}
+	}
+
+	/**
+	 * Map
+	 */
+
+	public async resetMapSources() {
+		await this.isViewerReady();
+		UnityUtil.resetMapSources();
+	}
+
+	public async addMapSource(source) {
+		await this.isViewerReady();
+		UnityUtil.addMapSource(source);
+	}
+
+	public async removeMapSource(source) {
+		await this.isViewerReady();
+		UnityUtil.removeMapSource(source);
+	}
+
+	public async mapInitialise(surveyPoints) {
+		await this.isViewerReady();
+		UnityUtil.mapInitialise(surveyPoints);
+	}
+
+	public async mapStart() {
+		await this.isViewerReady();
+		UnityUtil.mapStart();
+	}
+
+	public async mapStop() {
+		await this.isViewerReady();
+		UnityUtil.mapStop();
+	}
+
+	/**
+	 * MultiSelect
+	 */
+
+	public async goToExtent() {
+		await this.isViewerReady();
+		this.showAll();
+	}
+
+	/**
+	 * Status
+	 */
+
+	public async getObjectsStatus({ teamspace, model } = { teamspace: '', model: '' }) {
+		await this.isViewerReady();
+		return this.getUnityObjectsStatus(teamspace, model);
+	}
+
+	public async getCurrentViewpoint({ teamspace, model }) {
+		return await this.getCurrentViewpointInfo(teamspace, model);
+	}
+
+	public async goToDefaultViewpoint() {
+		await this.isViewerReady();
+		return this.showAll();
+	}
+
+	public async setCamera({ position, up, view_dir, look_at, animate, rollerCoasterMode, account, model }) {
+		await this.isModelReady();
+		UnityUtil.setViewpoint(position, up, view_dir, look_at, animate !== undefined ? animate : true, rollerCoasterMode);
+	}
+
+	/**
+	 * Camera
+	 */
+
+	public topView() {
+		UnityUtil.topView();
+	}
+
+	public bottomView() {
+		return UnityUtil.bottomView();
+	}
+
+	public frontView() {
+		return UnityUtil.frontView();
+	}
+
+	public backView() {
+		return UnityUtil.backView();
+	}
+
+	public leftView() {
+		return UnityUtil.leftView();
+	}
+
+	public rightView() {
+		return UnityUtil.rightView();
+	}
+
+	public showAll() {
+		UnityUtil.resetCamera();
+	}
+
+	public centreToPoint(params) {
+		UnityUtil.centreToPoint(params);
+	}
+
+	/**
+	 * Controlling
+	 */
+
+	public pauseRendering() {
+		UnityUtil.pauseRendering();
+	}
+
+	public resumeRendering() {
+		UnityUtil.resumeRendering();
 	}
 
 	public setModelCache = (cache: boolean) => {
 		if (cache) {
-			this.viewer.enableCaching();
+			UnityUtil.enableCaching();
 		} else {
-			this.viewer.disableCaching();
+			UnityUtil.disableCaching();
 		}
 	}
 
-	public pauseRendering = () => {
-		this.viewer.pauseRendering();
+	public switchObjectVisibility(account, model, ids, visibility) {
+		UnityUtil.toggleVisibility(account, model, ids, visibility);
 	}
 
-	public resumeRendering = () => {
-		this.viewer.resumeRendering();
+	public startAreaSelect() {
+		UnityUtil.startAreaSelection();
+	}
+
+	public stopAreaSelect() {
+		UnityUtil.stopAreaSelection();
+	}
+
+	public setShadows = (type: string) => {
+		switch (type) {
+			case 'soft':
+				this.enableSoftShadows();
+				break;
+			case 'hard':
+				this.enableHardShadows();
+				break;
+			case 'none':
+				this.disableShadows();
+				break;
+		}
+	}
+
+	public enableSoftShadows() {
+		UnityUtil.enableSoftShadows();
+	}
+
+	public enableHardShadows() {
+		UnityUtil.enableHardShadows();
+	}
+
+	public disableShadows() {
+		UnityUtil.disableShadows();
+	}
+
+	/**
+	 * Clip
+	 */
+
+	public getNumPlanes() {
+		return this.numClips;
+	}
+
+	public clipBroadcast(clip) {
+		this.emit(VIEWER_EVENTS.CLIPPING_PLANE_BROADCAST, clip);
+	}
+
+	public numClipPlanesUpdated(nPlanes) {
+		this.numClips = nPlanes;
+		this.emit(VIEWER_EVENTS.UPDATE_NUM_CLIP, nPlanes);
+	}
+
+	public updateClippingPlanes({ clipPlanes, account, model }: any) {
+		UnityUtil.updateClippingPlanes(clipPlanes ? clipPlanes : [], false, account, model);
+	}
+
+	public startClip(isSingle: boolean) {
+		if (isSingle) {
+			this.startSingleClip();
+		} else {
+			this.startBoxClip();
+		}
+	}
+
+	public startBoxClip() {
+		UnityUtil.startBoxClip();
+	}
+
+	public startSingleClip() {
+		UnityUtil.startSingleClip();
+	}
+
+	public startClipEdit() {
+		UnityUtil.startClipEdit();
+	}
+
+	public stopClipEdit() {
+		UnityUtil.stopClipEdit();
 	}
 }
 
-export const Viewer = new ViewerService();
+export const Viewer = new ViewerService({});
 
 export const withViewer = (WrappedComponent) => (props) => (
-	<WrappedComponent viewer={Viewer} {...props} />
+		<WrappedComponent viewer={Viewer} {...props} />
 );
